EPS works fine, some work on various TIFF compression methods and colorspaces
1. Made TEPSGraphic compatible with old TTIFFGraphic, but should work
with new one as well.
.EPS files from test suite are working fine.

2. working on fax3 2D compression as well as all fax3 compression code
working on both x86 and x64. Still in progress.

3. Working on L and Luv support for TIFF, so far it determines
compression method and colorspace correctly but doesn't know what to do
with it...

4. Trying to make JPEG compression in TIFF files work. Strange thing
that JPEG.pas has wrong call conventions: register instead of cdecl. Now
it proceeds a bit further but fails on AV at address 0000000.